+++ samanthas_telegram_bot/conversation/auxil/callback_query_reply_sender.py
@@ -19,7 +19,7 @@ from samanthas_telegram_bot.data_structures.enums import (
     Role,
     UserDataReviewCategory,
 )
-from samanthas_telegram_bot.data_structures.helper_classes import AssessmentQuestion
+from samanthas_telegram_bot.data_structures.models_helper_classes import AssessmentQuestion
 
 logger = logging.getLogger(__name__)
