+++ samanthas_telegram_bot/data_structures/context_types.py
@@ -7,7 +7,7 @@ from telegram.ext import CallbackContext, ExtBot
 from samanthas_telegram_bot.conversation.auxil.load_phrases import load_phrases
 from samanthas_telegram_bot.data_structures.constants import Locale
 from samanthas_telegram_bot.data_structures.enums import AgeRangeType, ConversationMode, Role
-from samanthas_telegram_bot.data_structures.helper_classes import (
+from samanthas_telegram_bot.data_structures.models_helper_classes import (
     Assessment,
     AssessmentAnswer,
     DayAndTimeSlot,
